add global window interface with shopify and analytics
